+++ datadriven/src/sgpp/datadriven/operation/hash/OperationMPI/OperationGridBaseMPI.hpp
@@ -65,14 +65,14 @@ class MPIWorkerGridBase : virtual public MPIWorkerBase {
     // Store grid in integer array
     std::cout << "IN GridWorker cstr" << "\n";
     sgpp::base::GridStorage& gridStorage = grid.getStorage();
-    gridsize = gridStorage.getSize();
-    size_t dimensions = gridStorage.getDimension();
+    gridsize = static_cast<int>(gridStorage.getSize());
+    int dimensions = static_cast<int>(gridStorage.getDimension());
     gridpoints = new int[gridsize * 2 * dimensions];
     size_t pointscount = 0;
     for (int i = 0; i < gridsize; i++) {
       sgpp::base::HashGridPoint &point = gridStorage.getPoint(i);
       pointscount++;
-      for (size_t d = 0; d < dimensions; d++) {
+      for (int d = 0; d < dimensions; d++) {
         gridpoints[i * 2 * dimensions + 2 * d] = point.getIndex(d);
         gridpoints[i * 2 * dimensions + 2 * d + 1] = point.getLevel(d);
       }
